backend/requirements.txt

fastapi 
uvicorn 
sqlalchemy 
pydantic 
requests 
python-dotenv
psycopg2-binary 
python-jose[cryptography]
passlib
email-validator
python-multipart
tzdata
